Move and rename CommandListener
