+++ src/main/kotlin/me/aberrantfox/hotbot/commandframework/commands/EmbedCommands.kt
@@ -1,10 +1,9 @@
 package me.aberrantfox.hotbot.commandframework.commands
 
-import me.aberrantfox.hotbot.commandframework.ArgumentType
+import me.aberrantfox.hotbot.commandframework.parsing.ArgumentType
 import me.aberrantfox.hotbot.dsls.command.CommandSet
 import me.aberrantfox.hotbot.dsls.command.arg
 import me.aberrantfox.hotbot.dsls.command.commands
-import me.aberrantfox.hotbot.extensions.*
 import me.aberrantfox.hotbot.extensions.jda.fullName
 import me.aberrantfox.hotbot.extensions.stdlib.isBooleanValue
 import me.aberrantfox.hotbot.extensions.stdlib.isLong
