+++ src/main/kotlin/me/aberrantfox/hotbot/commandframework/commands/FunCommands.kt
@@ -2,7 +2,7 @@ package me.aberrantfox.hotbot.commandframework.commands
 
 import com.github.ricksbrown.cowsay.Cowsay
 import khttp.get as kget
-import me.aberrantfox.hotbot.commandframework.ArgumentType
+import me.aberrantfox.hotbot.commandframework.parsing.ArgumentType
 import me.aberrantfox.hotbot.dsls.command.CommandSet
 import me.aberrantfox.hotbot.dsls.command.commands
 import org.jsoup.Jsoup
