+++ src/main/kotlin/me/aberrantfox/hotbot/commandframework/commands/MacroCommands.kt
@@ -2,7 +2,7 @@ package me.aberrantfox.hotbot.commandframework.commands
 
 import com.google.common.reflect.TypeToken
 import com.google.gson.Gson
-import me.aberrantfox.hotbot.commandframework.ArgumentType
+import me.aberrantfox.hotbot.commandframework.parsing.ArgumentType
 import me.aberrantfox.hotbot.dsls.command.CommandSet
 import me.aberrantfox.hotbot.dsls.command.commands
 import me.aberrantfox.hotbot.services.CommandRecommender
