+++ src/main/kotlin/me/aberrantfox/hotbot/commandframework/commands/SecurityCommands.kt
@@ -1,6 +1,6 @@
 package me.aberrantfox.hotbot.commandframework.commands
 
-import me.aberrantfox.hotbot.commandframework.ArgumentType
+import me.aberrantfox.hotbot.commandframework.parsing.ArgumentType
 import me.aberrantfox.hotbot.dsls.command.CommandSet
 import me.aberrantfox.hotbot.dsls.command.commands
 import me.aberrantfox.hotbot.listeners.antispam.NewPlayers
