+++ src/main/kotlin/me/aberrantfox/hotbot/commandframework/commands/UtilityCommands.kt
@@ -1,7 +1,7 @@
 package me.aberrantfox.hotbot.commandframework.commands
 
 import com.google.gson.Gson
-import me.aberrantfox.hotbot.commandframework.ArgumentType
+import me.aberrantfox.hotbot.commandframework.parsing.ArgumentType
 import me.aberrantfox.hotbot.dsls.command.CommandSet
 import me.aberrantfox.hotbot.database.savePermissions
 import me.aberrantfox.hotbot.dsls.command.CommandEvent
