+++ src/main/kotlin/me/aberrantfox/hotbot/dsls/command/CommandDSL.kt
@@ -1,6 +1,6 @@
 package me.aberrantfox.hotbot.dsls.command
 
-import me.aberrantfox.hotbot.commandframework.ArgumentType
+import me.aberrantfox.hotbot.commandframework.parsing.ArgumentType
 import me.aberrantfox.hotbot.extensions.stdlib.sanitiseMentions
 import me.aberrantfox.hotbot.logging.BotLogger
 import me.aberrantfox.hotbot.logging.DefaultLogger
